+++ NEWS.md
@@ -1,4 +1,4 @@
-# tsibble (development version)
+# tsibble 1.2.0
 
 * Fixed `scan_gaps()` failing with data containing the `int` column name.
 * Removed `as_tibble.grouped_df()` to prevent masking `dplyr::as_tibble.grouped_df()`.